Detect OS, and location of sox. Also, safety.
Almost total rewrite.

We now detect Linux or macOS, and the relative location of Sox for the OS.

I've also added a variable to protect the innocent from deleting their data. By default, the script will now just report the files the script thinks are silent, versus actually deleting them.

And, I've added a lot more information to the output while it runs.